+++ src/suffix_amd.js
@@ -1,0 +1,9 @@
+/**
+ * API suffix JS for AMD exposition
+ * Created on 16/12/2015.
+ * @author Tony Georges - Smart/Origin
+ */
+
+
+  return osmb;
+}));